Updated paths
